fix(4181): preserve frontier reset race guards
Keep captured frontier tokens authoritative through redrive admission and move reset coordination out of the giant Discord root without raising its baseline.
