+++ ARCHITECTURE.md
@@ -623,6 +623,8 @@ src/
 │   │   │   ├── mod.rs
 │   │   │   ├── restart.rs
 │   │   │   └── shared.rs
+│   │   ├── relay_health/
+│   │   │   └── frontier.rs
 │   │   ├── relay_recovery/
 │   │   │   └── tests/
 │   │   │       └── circuit_breaker_apply.rs
