+++ src/services/discord/mod.rs
@@ -132,8 +132,10 @@ mod voice_sensitivity;
 #[path = "watchers/lifecycle_decision.rs"]
 mod watcher_lifecycle_decision;
 
-pub(in crate::services::discord) use delivery_lease_key::DeliveryLeaseKey;
 pub(crate) use meeting_orchestrator as meeting;
+pub(in crate::services::discord) use {
+    delivery_lease_key::DeliveryLeaseKey, relay_health::RelayFrontierToken,
+};
 // #3479 item-2: re-export the catch-up subsystem entry points referenced
 // outside the extracted cluster (`maybe_schedule_catch_up_retry_after_queue_drain`
 // here in mod.rs and `catch_up_missed_messages` in runtime_bootstrap recovery).
@@ -1454,29 +1456,11 @@ mod tmux_watcher_registry_restore_tests {
 
 /// Per-channel coordination for watcher-to-Discord relay emission.
 ///
-/// This state is **shared across watcher-handle replacements** (unlike
-/// `TmuxWatcherHandle`, which is recreated on watcher reattach). It keeps
-/// relay emission serialized if a stale outgoing watcher overlaps with its
-/// successor, and it exposes the confirmed-output watermark used by watcher
-/// stop checks.
-///
-/// Scope: intra-process only. Persisted dedupe across dcserver restarts is
-/// still handled by `InflightTurnState::last_watcher_relayed_offset` in the
-/// inflight JSON.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(in crate::services::discord) struct RelayFrontierToken {
-    pub(in crate::services::discord) reset_incarnation: u64,
-    pub(in crate::services::discord) committed_offset: u64,
-}
-
+/// Shared across watcher-handle replacements, this serializes overlapping
+/// outgoing/successor relay emission and exposes the confirmed-output watermark.
+/// Scope: intra-process only; restart-persistent dedupe remains in
+/// `InflightTurnState::last_watcher_relayed_offset`.
 pub(super) struct TmuxRelayCoord {
-    /// Serializes downward watermark resets. Offset advances remain lock-free;
-    /// a reset publishes the new offset before it increments the incarnation.
-    reset_lock: std::sync::Mutex<()>,
-    /// Incremented after every committed-frontier decrease. Consumers pair this
-    /// with `confirmed_end_offset` through `frontier_token` so a stale coordinate
-    /// cannot be mistaken for the current JSONL incarnation.
-    reset_incarnation: std::sync::atomic::AtomicU64,
     /// Non-zero while some watcher instance is actively emitting a relay for
     /// this channel. Holds the `data_start_offset` of the in-progress emission.
     /// Acquired via `compare_exchange(0, offset)` — only one watcher can
@@ -1500,6 +1484,7 @@ pub(super) struct TmuxRelayCoord {
     /// solely because another watcher advanced this watermark; only the
     /// no-inflight wake/idle paths gate on it.
     pub(super) confirmed_end_offset: Arc<std::sync::atomic::AtomicU64>,
+    pub(in crate::services::discord) reset_state: std::sync::Mutex<u64>,
     /// Wall-clock timestamp (ms since epoch) of the most recent confirmed
     /// relay. 0 = no confirmed relay observed yet. Read by the
     /// `watcher-state` observability endpoint (#964). Monotonic is NOT
@@ -1539,10 +1524,9 @@ pub(super) struct TmuxRelayCoord {
 impl TmuxRelayCoord {
     pub(super) fn new(channel_id: ChannelId) -> Self {
         Self {
-            reset_lock: std::sync::Mutex::new(()),
-            reset_incarnation: std::sync::atomic::AtomicU64::new(0),
             relay_slot: Arc::new(std::sync::atomic::AtomicU64::new(0)),
             confirmed_end_offset: Arc::new(std::sync::atomic::AtomicU64::new(0)),
+            reset_state: std::sync::Mutex::new(0),
             last_relay_ts_ms: Arc::new(std::sync::atomic::AtomicI64::new(0)),
             reconnect_count: Arc::new(std::sync::atomic::AtomicU64::new(0)),
             confirmed_end_generation_mtime_ns: Arc::new(std::sync::atomic::AtomicI64::new(0)),
@@ -1552,43 +1536,6 @@ impl TmuxRelayCoord {
 
     pub(super) fn note_relay_progress_heartbeat(&self, now_ms: i64) {
         self.last_relay_ts_ms.store(now_ms, Ordering::Release);
-    }
-
-    pub(in crate::services::discord) fn frontier_token(&self) -> RelayFrontierToken {
-        let _reset = self
-            .reset_lock
-            .lock()
-            .unwrap_or_else(|error| error.into_inner());
-        RelayFrontierToken {
-            reset_incarnation: self.reset_incarnation.load(Ordering::Acquire),
-            committed_offset: self.confirmed_end_offset.load(Ordering::Acquire),
-        }
-    }
-
-    pub(in crate::services::discord) fn reset_confirmed_frontier(
-        &self,
-        expected_offset: u64,
-        new_offset: u64,
-    ) -> bool {
-        debug_assert!(new_offset < expected_offset);
-        let _reset = self
-            .reset_lock
-            .lock()
-            .unwrap_or_else(|error| error.into_inner());
-        if self
-            .confirmed_end_offset
-            .compare_exchange(
-                expected_offset,
-                new_offset,
-                Ordering::AcqRel,
-                Ordering::Acquire,
-            )
-            .is_err()
-        {
-            return false;
-        }
-        self.reset_incarnation.fetch_add(1, Ordering::AcqRel);
-        true
     }
 }
 
@@ -1606,20 +1553,6 @@ mod relay_coord_tests {
         coord.note_relay_progress_heartbeat(1_500);
         assert_eq!(coord.last_relay_ts_ms.load(Ordering::Acquire), 1_500);
         assert_eq!(coord.confirmed_end_offset.load(Ordering::Acquire), 0);
-    }
-
-    #[test]
-    fn reset_frontier_publishes_a_new_incarnation_token() {
-        let coord = TmuxRelayCoord::new(ChannelId::new(4_181));
-        coord.confirmed_end_offset.store(100, Ordering::Release);
-        let high = coord.frontier_token();
-        assert!(coord.reset_confirmed_frontier(100, 40));
-        let low = coord.frontier_token();
-
-        assert_eq!(high.committed_offset, 100);
-        assert_eq!(low.committed_offset, 40);
-        assert!(low.reset_incarnation > high.reset_incarnation);
-        assert_ne!(high, low, "a reset must invalidate stale redrive tokens");
     }
 }
 
@@ -2455,21 +2388,6 @@ impl SharedData {
         self.tmux_relay_coord(channel_id)
             .confirmed_end_offset
             .load(Ordering::Acquire)
-    }
-
-    pub(in crate::services::discord) fn relay_frontier_token(
-        &self,
-        channel_id: ChannelId,
-    ) -> RelayFrontierToken {
-        self.tmux_relay_coord(channel_id).frontier_token()
-    }
-
-    pub(in crate::services::discord) fn relay_frontier_token_is_current(
-        &self,
-        channel_id: ChannelId,
-        token: RelayFrontierToken,
-    ) -> bool {
-        self.relay_frontier_token(channel_id) == token
     }
 
     /// #4181: `true` while some watcher is actively emitting a relay for this
